vp8: Change type of stride parameters to ptrdiff_t
ptrdiff_t is the correct type for array strides and similar.

+++ libavcodec/vp8.h
@@ -143,8 +143,8 @@ typedef struct VP8Context {
 
     uint16_t mb_width;   /* number of horizontal MB */
     uint16_t mb_height;  /* number of vertical MB */
-    int linesize;
-    int uvlinesize;
+    ptrdiff_t linesize;
+    ptrdiff_t uvlinesize;
 
     uint8_t keyframe;
     uint8_t deblock_filter;
